requirements.txt
Django==5.1.3
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
celery==5.4.0
django-celery-beat==2.7.0
rest_framework_simplejwt